ticket_parser/requirements.txt
markdown-it-py>=3.0.0
jsonschema>=4.20.0
pytest>=7.4.0
